Fix pause before final remainder part
