+++ lib/upload.js
@@ -288,12 +288,11 @@ class BaseUpload {
     this._parallelUploadUrls = new Array(parts.length)
 
     let firstChunk = null
-    let currentChunk = null
+    let partsChunks = null
     if (staggerPercent !== null) {
       // Generate a promise for each chunk that will be resolved when the respective
       // upload is completed.
-      firstChunk = buildChunkStaggers(totalSize, parts.length, chunkSize)
-      currentChunk = firstChunk
+      [partsChunks, firstChunk] = buildChunkStaggers(totalSize, parts.length, chunkSize)
     }
 
     const uploads = parts.map((part, index) => {
@@ -357,7 +356,7 @@ class BaseUpload {
               const chunk = upload._currentChunk
               const chunkProgress = (newPartProgress - (chunkSize * chunk.indexInPart))
               const chunkPercent = (chunkProgress / chunkSize) * 100
-              console.log('part:', index + 1, 'chunk:', chunk.indexInPart, '| chunk percent:', chunkPercent)
+              console.log('part:', index + 1, 'chunk:', chunk.indexInPart + 1, '| chunk percent:', chunkPercent)
               const nextChunk = chunk.nextChunkStagger
 
               if (nextChunk !== null && !nextChunk.resolved && chunkPercent >= staggerPercent) {
@@ -373,13 +372,12 @@ class BaseUpload {
 
           let promise = Promise.resolve()
           if (staggerPercent !== null) {
+            const currentChunk = partsChunks[index][0]
+
             // Wait for the initial stagger for this chunk before starting the part.
             promise = currentChunk.promise
 
             upload._currentChunk = currentChunk
-            // Update the current chunk so that the next part starts at the right
-            // chunk.
-            currentChunk = currentChunk.nextChunkStagger
           }
           promise.then(() => upload.start())
 
@@ -1112,7 +1110,6 @@ function buildChunkStaggers (totalSize, partCount, chunkSize) {
   }
   // The leftover size that must go into the last part.
   const hasLeftover = (totalSize % chunkSize) > 0
-  let lastChunk = null
 
   function newChunk (index) {
     const partChunks = partsChunks[index]
@@ -1135,12 +1132,6 @@ function buildChunkStaggers (totalSize, partCount, chunkSize) {
 
     // Push chunk to part.
     partChunks.push(chunk)
-
-    // Last chunk in stagger order should point to this one.
-    if (lastChunk !== null) {
-      lastChunk.nextChunkStagger = chunk
-    }
-    lastChunk = chunk
   }
 
   // Assign chunks to parts in order, looping back to the beginning if we get to
@@ -1165,7 +1156,35 @@ function buildChunkStaggers (totalSize, partCount, chunkSize) {
     newChunk(lastIndex)
   }
 
-  return partsChunks[0][0]
+  // Now assign the stagger order. We start at the part with the most chunks, so
+  // we don't end up with any final pauses.
+  let longestPartIndex = null
+  let longestPartLength = 0
+  for (let i = 0; i < partsChunks.length; i++) {
+    if (longestPartIndex === null || partsChunks[i].length > longestPartLength) {
+      longestPartIndex = i
+      longestPartLength = partsChunks[i].length
+    }
+  }
+  // Get the chunk at which to start uploading.
+  const firstChunk = partsChunks[longestPartIndex][0]
+  // Build the stagger sequence.
+  let lastChunk = null
+  for (let i = 0; i < longestPartLength; i++) {
+    for (let j = longestPartIndex; j < longestPartIndex + partCount; j++) {
+      const index = j % partCount
+      if (partsChunks[index].length > i) {
+        const chunk = partsChunks[index][i]
+        // Last chunk in stagger order should point to this one.
+        if (lastChunk !== null) {
+          lastChunk.nextChunkStagger = chunk
+        }
+        lastChunk = chunk
+      }
+    }
+  }
+
+  return [partsChunks, firstChunk]
 }
 
 BaseUpload.defaultOptions = defaultOptions
