+++ test/spec/browser-index.js
@@ -36,5 +36,6 @@ require('./helpers/puppeteer/reporter')
 require('./test-common')
 require('./test-browser-specific')
 require('./test-parallel-uploads')
+require('./test-parallel-stagger-uploads')
 require('./test-terminate')
 require('./test-end-to-end')
