+++ test/spec/node-index.js
@@ -1,5 +1,6 @@
 require('./test-common')
 require('./test-node-specific')
 require('./test-parallel-uploads')
+require('./test-parallel-stagger-uploads')
 require('./test-terminate')
 require('./test-end-to-end')
